requirements.txt
aiohttp
imageio
numpy
Pillow
aiosqlite